feat(v18)!: retire full/ — libpg-query pg18+ ships the full API (13-17 stay slim)
