+++ .github/workflows/build-wasm-no-docker.yaml
@@ -42,15 +42,15 @@ jobs:
           ./emsdk install 3.1.59
           ./emsdk activate 3.1.59
           source ./emsdk_env.sh
-        working-directory: full/18
+        working-directory: versions/18
       - name: Build with Emscripten 🏗
         run: |
           source ./emsdk/emsdk_env.sh
           emmake make
           emmake make build
-        working-directory: full/18
+        working-directory: versions/18
       - name: Archive production artifacts 🏛
         uses: actions/upload-artifact@v4
         with:
           name: wasm-artifacts
-          path: full/18/wasm
+          path: versions/18/wasm
