+++ .github/workflows/ci.yml
@@ -13,8 +13,6 @@ jobs:
     strategy:
       matrix:
         package: 
-          - { name: 'full17', path: 'full/17', version: '17' }
-          - { name: 'full18', path: 'full/18', version: '18' }
           - { name: 'v13', path: 'versions/13', version: '13' }
           - { name: 'v14', path: 'versions/14', version: '14' }
           - { name: 'v15', path: 'versions/15', version: '15' }
@@ -79,8 +77,6 @@ jobs:
       matrix:
         os: [ubuntu-latest, macos-latest, windows-latest]
         package: 
-          - { name: 'full17', path: 'full/17', version: '17' }
-          - { name: 'full18', path: 'full/18', version: '18' }
           - { name: 'v13', path: 'versions/13', version: '13' }
           - { name: 'v14', path: 'versions/14', version: '14' }
           - { name: 'v15', path: 'versions/15', version: '15' }
