--- full/.npmignore
@@ -1,18 +1,0 @@
-# project files
-test
-.gitignore
-package.json
-build
-
-*.log
-npm-debug.log*
-
-# Dependency directories
-node_modules
-
-libpg_query/*
-
-# npm package lock
-package-lock.json
-yarn.lock
-
