--- full/17/tsconfig.esm.json
@@ -1,9 +1,0 @@
-{
-    "extends": "./tsconfig.json",
-    "compilerOptions": {
-      "module": "es2022",
-      "rootDir": "src/",
-      "declaration": false,
-      "outDir": "esm/"
-    }
-  }
