--- full/18/package.json
@@ -1,55 +1,0 @@
-{
-  "name": "@libpg-query/full18",
-  "version": "18.0.0",
-  "description": "The real PostgreSQL query parser",
-  "homepage": "https://github.com/constructive-io/libpg-query-node",
-  "main": "./wasm/index.cjs",
-  "module": "./wasm/index.js",
-  "typings": "./wasm/index.d.ts",
-  "publishConfig": {
-    "access": "public"
-  },
-  "x-publish": {
-    "publishName": "@libpg-query/parser",
-    "pgVersion": "18",
-    "distTag": "pg18",
-    "libpgQueryTag": "18-constructive"
-  },
-  "files": [
-    "wasm/*"
-  ],
-  "scripts": {
-    "clean": "pnpm wasm:clean && rimraf cjs esm",
-    "build:js": "node scripts/build.js",
-    "build": "pnpm clean; pnpm wasm:build; pnpm build:js",
-    "publish:pkg": "node ../../scripts/publish-single-version.js",
-    "wasm:make": "docker run --rm -v $(pwd):/src -u $(id -u):$(id -g) emscripten/emsdk emmake make",
-    "wasm:build": "pnpm wasm:make build",
-    "wasm:rebuild": "pnpm wasm:make rebuild",
-    "wasm:clean": "pnpm wasm:make clean",
-    "wasm:clean-cache": "pnpm wasm:make clean-cache",
-    "test": "node --test test/parsing.test.js test/pg18.test.js test/fingerprint.test.js test/normalize.test.js test/plpgsql.test.js test/scan.test.js test/errors.test.js",
-    "yamlize": "node ./scripts/yamlize.js"
-  },
-  "author": "Constructive <developers@constructive.io>",
-  "license": "MIT",
-  "repository": {
-    "type": "git",
-    "url": "git://github.com/constructive-io/libpg-query-node.git"
-  },
-  "devDependencies": {
-    "@yamlize/cli": "^0.8.0"
-  },
-  "dependencies": {
-    "@pgsql/types": "^18.0.0"
-  },
-  "keywords": [
-    "sql",
-    "postgres",
-    "postgresql",
-    "pg",
-    "query",
-    "plpgsql",
-    "database"
-  ]
-}
