+++ pnpm-workspace.yaml
@@ -1,7 +1,5 @@
 packages:
   - 'parser'
-  - 'full/18'
-  - 'full/17'
   - 'versions/18'
   - 'versions/17' 
   - 'versions/16'
