+++ scripts/analyze-sizes.js
@@ -74,10 +74,7 @@ function analyzePackage(packagePath, packageName) {
 // Get all version packages dynamically
 function getVersionPackages() {
   const versionsDir = './versions';
-  const packages = [
-    { path: './full/17', name: 'full/17 (Full)', version: 'original' },
-    { path: './full/18', name: 'full/18 (Full)', version: 'original' }
-  ];
+  const packages = [];
 
   if (fs.existsSync(versionsDir)) {
     const versions = fs.readdirSync(versionsDir)
