+++ templates/full/index.ts
