+++ templates/full/wasm_wrapper.c
